Leaning more heavily on GPT4 over time.
This update introduces a new Director which tries to determine if the question/objective can be quickly answered from previous conversation or memory. If not, then the ReAct agent is used.

Additionally, the ActionEvaluator was updated to use GPT4. The cost is real, but without it, the action was rarely improved. Needs more analysis.